Refactor functions to patternMatching.ts
